#19 [feat] useModal 훅 분리
